LayoutTests/canvas/philip/tests/2d.imageData.create2.nonfinite-expected.txt

< [index] >
2d.imageData.create2.nonfinite
createImageData() throws NOT_SUPPORTED_ERR if arguments are not finite
References: 2d.imageData.getcreate.nonfinite Bugzilla
Actual output:
Passed
